+++ app/src/test/java/dop/chapter04/Listing4_1.java
@@ -1,13 +1,8 @@
 package dop.chapter04;
 
-
-import org.junit.jupiter.api.Assertions;
-import org.junit.jupiter.api.Test;
-
-import java.time.Instant;
 import java.util.List;
-import java.util.Map;
-import java.util.function.Function;
+
+import org.junit.jupiter.api.Test;
 
 /**
  * Chapter 4 builds on top of chapter 3's exploration of
@@ -26,33 +21,34 @@ import java.util.function.Function;
  */
 public class Listing4_1 {
 
-    /**
-     * ───────────────────────────────────────────────────────
-     * Listings 4.1
-     * ───────────────────────────────────────────────────────
-     * We're going to model a checklist! How could anyone possibly
-     * screw up making a checklist? I'm pretty sure I accidentally
-     * found all the ways.
-     *
-     * We might start with the "obvious" approach:
-     * ───────────────────────────────────────────────────────
-     */
-    @Test
-    public void example() {
-        //             ┌── A named "thing to do"
-        //             ▼
-        record Step(String name, boolean isComplete) {
-        }
-        //                           ▲
-        //                           └── And it's either done or not
-        record Checklist(List<Step> steps) {
-        }
-        //          ▲
-        //          └─ A checklist is just a collection of steps.
-        //             If all the steps are complete. The checklist is complete.
-        // Easy Peasy?
-        //
-        // Done?
+  /**
+   * ───────────────────────────────────────────────────────
+   * Listings 4.1
+   * ───────────────────────────────────────────────────────
+   * We're going to model a checklist! How could anyone possibly
+   * screw up making a checklist? I'm pretty sure I accidentally
+   * found all the ways.
+   *
+   * We might start with the "obvious" approach:
+   * ───────────────────────────────────────────────────────
+   */
+  @Test
+  public void example() {
+    //             ┌── A named "thing to do"
+    //             ▼
+    record Step(String name, boolean isComplete) {
     }
+    //                           ▲
+    //                           └── And it's either done or not
+    record Checklist(List<Step> steps) {
+    }
+    //          ▲
+    //          └─ A checklist is just a collection of steps.
+    //             If all the steps are complete. The checklist is complete.
+    // Easy Peasy?
+    //
+    // Done?
+  }
+
 }
