+++ app/src/test/java/dop/chapter04/Listing4_13.java
@@ -4,26 +4,28 @@ import java.time.Instant;
 import java.util.List;
 
 public class Listing4_13 {
-    /**
-     * ───────────────────────────────────────────────────────
-     * Listing 4.13
-     * ───────────────────────────────────────────────────────
-     * The finished data model
-     * ───────────────────────────────────────────────────────
-     */
-    record Step(String name){}
 
-    record Template(String name, List<Step> steps){}
+  /**
+   * ───────────────────────────────────────────────────────
+   * Listing 4.13
+   * ───────────────────────────────────────────────────────
+   * The finished data model
+   * ───────────────────────────────────────────────────────
+   */
+  record Step(String name){}
 
-    record Instance(
-        String name,
-        Instant date,
-        Template template
-    ){}
+  record Template(String name, List<Step> steps){}
 
-    record Status(
-        Template template,
-        Step step,
-        boolean isCompleted
-    ){}
+  record Instance(
+    String name,
+    Instant date,
+    Template template
+  ) {}
+
+  record Status(
+    Template template,
+    Step step,
+    boolean isCompleted
+  ) {}
+
 }
