+++ app/src/test/java/dop/chapter04/Listing4_35.java
@@ -2,13 +2,13 @@ package dop.chapter04;
 
 public class Listing4_35 {
 
-    /**
-     * ───────────────────────────────────────────────────────
-     * Listing 4.35
-     * ───────────────────────────────────────────────────────
-     * Our closed definition of ChecklistStatus
-     * ───────────────────────────────────────────────────────
-     */
+  /**
+   * ───────────────────────────────────────────────────────
+   * Listing 4.35
+   * ───────────────────────────────────────────────────────
+   * Our closed definition of ChecklistStatus
+   * ───────────────────────────────────────────────────────
+   */
   /*
   Status ==
       NotStarted
